requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==3.0.1
click==8.1.3
coreapi==2.3.3
coreschema==0.0.4
coverage==7.1.0
cryptography==39.0.0
Deprecated==1.2.13
distlib==0.3.6
Django==4.1.5
django-cors-headers==3.13.0
django-filter==22.1
django-oauth-toolkit==2.2.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
drf-nested-routers==0.93.4
drf-yasg==1.21.4
filelock==3.9.0
gunicorn==20.1.0
h11==0.14.0
identify==2.5.15
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jwcrypto==1.4.2
Markdown==3.4.1
MarkupSafe==2.1.2
networkx==3.0
nodeenv==1.7.0
oauthlib==3.2.2
packaging==23.0
platformdirs==2.6.2
pre-commit==3.0.0
psycopg2-binary==2.9.5
pycparser==2.21
python-dotenv==0.21.1
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.14
uvicorn==0.20.0
virtualenv==20.17.1
wrapt==1.14.1
